Add files via upload
Added an option to select Sleep (hdparm -Y), rather than suspend (hdparm -y), for cases where lowest power state is needed. May cause drives to vanish or mess with some RAID controllers. The only option is complete removal, and re-installation of the drive (if hotswappable) or restarting the system, in this case.